cleaner for parameters/predictions

+++ dev/clean_subjects_parameters.py
@@ -11,6 +11,8 @@ if __name__ == '__main__':
     for subject in os.listdir('./Source/Subjects'):
         nn_path = os.path.join('./Source/Subjects', subject, 'LFCNN')
         subject_params = os.path.join(nn_path, 'Parameters')
+        if os.path.exists(subject_params):
+            shutil.rmtree(subject_params)
         subject_preds = subject_params = os.path.join(nn_path, 'Predictions')
-        shutil.rmtree(subject_params)
-        shutil.rmtree(subject_preds)
+        if os.path.exists(subject_preds):
+            shutil.rmtree(subject_preds)
